+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2022-03-03
+### Added
+- Optionally compare two separate blind test runs and compute comparison metrics with `compare_results.py`
+
 ## 2020-11-12
 ### Changed
 - Pass Watson Assistant system_settings to test workspaces